\name{print.seedstack}

\alias{print.seedstack}

\title{
Print stack object.
}

\description{
  Print method for seedstack object.
}

\usage{
\method{print}{seedstack}(x, \dots)
}

\arguments{
  \item{x}{A data frame to print pretty.}
  \item{\dots}{Not used.}
}

\author{
  Kevin Wright
}

\keyword{models}

