\name{rseedcalc-package}
\alias{rseedcalc-package}
\alias{rseedcalc}
\docType{package}
\title{
Estimation of the proportion of genetically modified stacked seeds in seedlots
}
\description{
  Estimate the percentage of seeds in a seedlot that contain stacks
  of genetically modified traits.
}
\details{
  The main functions for this package are \link{stack2} and
  \link{stack3}.
  
  For a complete list of functions, use \code{library(help="utils")}
}
\author{
  Kevin Wright, Jean-Louis Laffont
  
  Maintainer: Kevin Wright <kw.stat@gmail.com>
}

\keyword{ package }
